\name{plotHMMmix}
\alias{plotHMMmix}

\title{plotHMMmix}
\description{
Specific plot function. Plot 4 graphs: the initial fit with K components, the values of BIC and ICLs criterion, the classification with D clusters and the classification with the number of clusters chosen by ICLs
}
\usage{
plotHMMmix(x)
}


\arguments{
  \item{x}{The object returned by the HMMmix function }

}






